requirements.txt
Django==3.1
django-extensions==3.1.5
djangorestframework==3.13.1
celery==5.2.3
redis==4.2.2
django-redis==5.2.0
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0
django-elasticsearch-dsl==8.0
google-api-python-client==2.121.0
